[MRA10-200] ETL error - Jasper Report Output (POI version problem)

+++ engine/src/org/pentaho/di/trans/steps/excelinput/staxpoi/StaxPoiSheet.java
@@ -278,7 +278,7 @@ public class StaxPoiSheet implements KSheet {
       CTXf cellXf = styles.getCellXfAt( styleIdx );
       if ( cellXf != null ) {
         // need id for builtin types, format if custom
-        int formatId = (int) cellXf.getNumFmtId();
+        short formatId = (short) cellXf.getNumFmtId();
         String format = styles.getNumberFormatAt( formatId );
         return DateUtil.isADateFormat( formatId, format );
       }
